move stat classes to separate modules per type, moved maintenance stat gathering into repository_host and updated specs
