+++ spec/models/maintenance_stats/average_commit_date_stats_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-describe MaintenanceStats::Stats::AverageCommitDate do
+describe MaintenanceStats::Stats::Github::AverageCommitDate do
   let!(:auth_token) { create(:auth_token) }
   let(:client) { auth_token.v4_github_client }
   let(:query_klass) { MaintenanceStats::Queries::FullRepoQuery.new(client) }
