+++ spec/models/maintenance_stats/contributor_count_stats_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-describe MaintenanceStats::Stats::V3ContributorCountStats do
+describe MaintenanceStats::Stats::Github::V3ContributorCountStats do
   let!(:auth_token) { create(:auth_token) }
   let(:client) { AuthToken.client }
   let(:query_klass) { MaintenanceStats::Queries::RepositoryContributorStatsQuery.new(client) }
