+++ spec/models/maintenance_stats/issues_stats_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-describe MaintenanceStats::Stats::IssueRates do
+describe MaintenanceStats::Stats::Github::IssueRates do
   let!(:auth_token) { create(:auth_token) }
   let(:client) { auth_token.v4_github_client }
   let(:query_klass) { MaintenanceStats::Queries::FullRepoQuery.new(client) }
